* feature/multi-cf-instance: (47 commits)
  Remote connection fixes
  Couple small style fixes
  Add test for remotely connected namespace services
  Simplify isRemoteConnectBinding
  Add support for creating remote connection buckets
  Don’t add remote connection keys to instance references
  Clean up warnings
  Clean up test warnings & optional usage
  Clean errors & improve error logging
  Add support for namespace deletes with remote connections
  Add support for bucket deletes with remote connections
  Move change reference check into EcsServiceInstanceService
  Prevent namespace plan change when remote connections exist
  Prevent bucket plan change when remote connections exist
  Add remote connection workflows & update tests
  Further refactor to extract remote-connection workflow
  Add support for “remote connection key” credentials in bindings
  Bump powermock version
  Refactor EcsServiceBindingService with workflow class
  Update README.md
  ...

+++ src/main/java/com/emc/ecs/cloudfoundry/broker/service/EcsServiceInstanceService.java
@@ -6,10 +6,11 @@ import com.emc.ecs.cloudfoundry.broker.model.PlanProxy;
 import com.emc.ecs.cloudfoundry.broker.model.ServiceDefinitionProxy;
 import com.emc.ecs.cloudfoundry.broker.repository.ServiceInstance;
 import com.emc.ecs.cloudfoundry.broker.repository.ServiceInstanceRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.servicebroker.exception.ServiceBrokerException;
 import org.springframework.cloud.servicebroker.exception.ServiceInstanceDoesNotExistException;
-import org.springframework.cloud.servicebroker.exception.ServiceInstanceExistsException;
 import org.springframework.cloud.servicebroker.exception.ServiceInstanceUpdateNotSupportedException;
 import org.springframework.cloud.servicebroker.model.*;
 import org.springframework.cloud.servicebroker.service.ServiceInstanceService;
@@ -17,14 +18,19 @@ import org.springframework.stereotype.Service;
 
 import java.net.URISyntaxException;
 import java.util.Map;
-import java.util.Optional;
+
+import static java.lang.String.format;
 
 @Service
 public class EcsServiceInstanceService implements ServiceInstanceService {
+
+    private static final Logger logger = LoggerFactory.getLogger(EcsServiceInstanceService.class);
+
     private static final String NO_SERVICE_MATCHING_TYPE = "No service matching type: ";
     private static final String NAMESPACE = "namespace";
     private static final String BUCKET = "bucket";
     private static final String SERVICE_TYPE = "service-type";
+    private static final Logger LOG = LoggerFactory.getLogger(EcsServiceInstanceService.class);
 
     @Autowired
     private EcsService ecs;
@@ -38,59 +44,50 @@ public class EcsServiceInstanceService implements ServiceInstanceService {
     }
 
     @Override
-    public CreateServiceInstanceResponse createServiceInstance(
-            CreateServiceInstanceRequest request)
-            throws ServiceInstanceExistsException, ServiceBrokerException {
-
-        ServiceInstance instance = new ServiceInstance(request);
+    public CreateServiceInstanceResponse createServiceInstance(CreateServiceInstanceRequest request) {
         String serviceInstanceId = request.getServiceInstanceId();
         String serviceDefinitionId = request.getServiceDefinitionId();
         String planId = request.getPlanId();
+
+        logger.info(format("Creating service instance %s", serviceInstanceId));
+
         try {
             ServiceDefinitionProxy service = ecs
                     .lookupServiceDefinition(serviceDefinitionId);
             PlanProxy plan = service.findPlan(planId);
-            String serviceType = (String) service.getServiceSettings()
-                    .get(SERVICE_TYPE);
-            if (BUCKET.equals(serviceType)) {
-                ecs.createBucket(serviceInstanceId, service, plan,
-                        Optional.ofNullable(request.getParameters()));
-            } else if (NAMESPACE.equals(serviceType)) {
-                ecs.createNamespace(serviceInstanceId, service, plan,
-                        Optional.ofNullable(request.getParameters()));
-            } else {
-                throw new EcsManagementClientException(
-                        NO_SERVICE_MATCHING_TYPE + serviceType);
-            }
+            InstanceWorkflow workflow = getWorkflow(request)
+                    .withCreateRequest(request);
+
+            LOG.info("creating service instance");
+            ServiceInstance instance = workflow.create(serviceInstanceId, service, plan, request.getParameters());
+
+            LOG.info("saving instance...");
             repository.save(instance);
+
             return new CreateServiceInstanceResponse();
         } catch (Exception e) {
+            logger.error(format("Unexpected error creating service %s", serviceInstanceId), e);
             throw new ServiceBrokerException(e);
         }
     }
 
     @Override
-    public DeleteServiceInstanceResponse deleteServiceInstance(
-            DeleteServiceInstanceRequest request)
-            throws ServiceBrokerException {
+    public DeleteServiceInstanceResponse deleteServiceInstance(DeleteServiceInstanceRequest request) {
 
         String serviceInstanceId = request.getServiceInstanceId();
         String serviceDefinitionId = request.getServiceDefinitionId();
         try {
             ServiceDefinitionProxy service = ecs
                     .lookupServiceDefinition(serviceDefinitionId);
+            InstanceWorkflow workflow = getWorkflow(service)
+                    .withDeleteRequest(request);
 
-            String serviceType = (String) service.getServiceSettings()
-                    .get(SERVICE_TYPE);
-            if (BUCKET.equals(serviceType)) {
-                ecs.deleteBucket(serviceInstanceId);
-            } else if (NAMESPACE.equals(serviceType)) {
-                ecs.deleteNamespace(serviceInstanceId);
-            } else {
-                throw new EcsManagementClientException(
-                        NO_SERVICE_MATCHING_TYPE + serviceType);
-            }
+            LOG.info("deleting service instance");
+            workflow.delete(serviceInstanceId);
+
+            LOG.info("removing instance from repo");
             repository.delete(serviceInstanceId);
+
             return new DeleteServiceInstanceResponse();
         } catch (Exception e) {
             throw new ServiceBrokerException(e);
@@ -98,16 +95,10 @@ public class EcsServiceInstanceService implements ServiceInstanceService {
     }
 
     @Override
-    public UpdateServiceInstanceResponse updateServiceInstance(
-            UpdateServiceInstanceRequest request)
-            throws ServiceInstanceUpdateNotSupportedException,
-            ServiceBrokerException, ServiceInstanceDoesNotExistException {
-
+    public UpdateServiceInstanceResponse updateServiceInstance(UpdateServiceInstanceRequest request) {
         String serviceInstanceId = request.getServiceInstanceId();
         String serviceDefinitionId = request.getServiceDefinitionId();
         String planId = request.getPlanId();
-        Map<String, Object> params = request.getParameters();
-
         try {
             ServiceDefinitionProxy service = ecs
                     .lookupServiceDefinition(serviceDefinitionId);
@@ -115,27 +106,48 @@ public class EcsServiceInstanceService implements ServiceInstanceService {
             if (instance == null)
                 throw new ServiceInstanceDoesNotExistException(
                         serviceInstanceId);
+            if (instance.getReferences().size() > 1)
+                throw new ServiceInstanceUpdateNotSupportedException(
+                        "Cannot change plan of service instance with remote references");
 
-            String serviceType = (String) service.getServiceSettings()
-                    .get(SERVICE_TYPE);
-            PlanProxy plan = service.findPlan(planId);
-            if (BUCKET.equals(serviceType)) {
-                ecs.changeBucketPlan(serviceInstanceId, service, plan,
-                        Optional.ofNullable(params));
-            } else if (NAMESPACE.equals(serviceType)) {
-                ecs.changeNamespacePlan(serviceInstanceId, service, plan,
-                        params);
-            } else {
-                throw new EcsManagementClientException(
-                        NO_SERVICE_MATCHING_TYPE + serviceType);
-            }
+            InstanceWorkflow workflow = getWorkflow(service);
+            LOG.info("changing instance plan");
+            workflow.changePlan(serviceInstanceId, service, service.findPlan(planId), request.getParameters());
 
+            LOG.info("updating service in repo");
             repository.delete(serviceInstanceId);
             instance.update(request);
             repository.save(instance);
             return new UpdateServiceInstanceResponse();
         } catch (Exception e) {
             throw new ServiceBrokerException(e);
+        }
+    }
+
+    private InstanceWorkflow getWorkflow(CreateServiceInstanceRequest createRequest)
+            throws EcsManagementClientException {
+        if (isRemoteConnection(createRequest))
+            return new RemoteConnectionInstanceWorkflow(repository, ecs);
+        ServiceDefinitionProxy service = ecs.lookupServiceDefinition(createRequest.getServiceDefinitionId());
+        return getWorkflow(service);
+    }
+
+    private boolean isRemoteConnection(CreateServiceInstanceRequest createRequest) {
+        Map<String, Object> parameters = createRequest.getParameters();
+        return parameters != null && parameters.containsKey("remote_connection");
+    }
+
+    private InstanceWorkflow getWorkflow(ServiceDefinitionProxy service)
+            throws EcsManagementClientException {
+        String serviceType = (String) service.getServiceSettings().get(SERVICE_TYPE);
+        switch (serviceType) {
+            case NAMESPACE:
+                return new NamespaceInstanceWorkflow(repository, ecs);
+            case BUCKET:
+                return new BucketInstanceWorkflow(repository, ecs);
+            default:
+                throw new ServiceBrokerException(NO_SERVICE_MATCHING_TYPE +
+                        serviceType);
         }
     }
